djangoapi/requirements.txt
cffi==1.8.3
cryptography==1.5.2
Django==1.10.2
django-cors-headers==1.2.2
django-debug-toolbar==1.6
django-extensions==1.7.4
django-rest-knox==2.2.0
djangorestframework==3.4.7
idna==2.1
markdown2==2.3.1
mysqlclient==1.3.9
pyasn1==0.1.9
pycparser==2.16
PyMySQL==0.7.9
pyOpenSSL==16.2.0
requests==2.11.1
six==1.10.0
sqlparse==0.2.1
Werkzeug==0.11.11
#django-pyodbc-azure==1.10.0.1
#pyodbc==3.0.10